IntegrationTest v1

+++ src/main/java/com/gftworkshopcatalog/api/dto/controllers/ProductController.java
@@ -14,6 +14,8 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
 
+import java.util.NoSuchElementException;
+
 @RestController
 @RequestMapping("/products")
 @Tag(name = "Products", description = "Everything about the products")
@@ -86,7 +88,7 @@ public class ProductController {
     })
     public ResponseEntity<?> updateProduct(
             @Parameter(description = "Product ID", required = true)
-            @PathVariable("id") long productId,
+            @PathVariable("id") Long productId,
             @RequestBody Product product) {
         Product updatedProduct = productService.updateProduct(productId, product);
         return ResponseEntity.ok(updatedProduct);
@@ -134,9 +136,8 @@ public class ProductController {
         }
     }
 
-
     @PatchMapping("/{productId}/stock")
-    public ResponseEntity<?> updateProductStock(@PathVariable("productId") long productId, @RequestParam("newStock") long newStock) {
+    public ResponseEntity<?> updateProductStock(@PathVariable("productId") long productId, @RequestParam("newStock") int newStock) {
         try {
             Product updatedProduct = productService.updateProductStock(productId, newStock);
             return ResponseEntity.ok(updatedProduct);
